BrowserDetect namespace
