+++ init.rb
@@ -1,2 +1,2 @@
-require File.expand_path(File.dirname(__FILE__)+'/lib/browser_detect/browser_detect')
+require 'browser_detect'
 ActionView::Base.send(:include, BrowserDetect)
